kafka-Setup/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
scapy==2.4.5
kafka-python==2.0.2
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
joblib==1.0.1
requests==2.26.0

